+++ src/main/java/crux/crux_server/config/login/oauth2/CustomOAuth2User.java
@@ -5,5 +5,5 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 public interface CustomOAuth2User extends OAuth2User {
     String getOauth2Id();
     String getRegistrationId();
-    Long getMemberId();
+    Integer getMemberId();
 }
